requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
django-environ==0.11.2
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
pydantic==2.10.6
pydantic-core==2.27.2
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
watchdog==4.0.2
